fix: remove broken features
